+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Anticarium_Web
 
+- ADD: First time setup script for Raspberry Pi to automatically configure WEB server
+- MODIFY: Python scripts in this repo now use ANTICARIUM_WEB_PATH environment variable to access this repo
+- MODIFY: anticarium_web.example now uses ANTICARIUM_WEB_PATH environment variable to access scripts provided by this repo 
+- MODIFY: anticarium_web.example now uses ANTICARIUM_SERVER_IP environment variable as server ip
 - ADD: Apache2 prefork module configuration, in 'apache2.conf', to process all requests in single thread
 - ADD: `apache2.conf.example` apache2 configuration example file
 - MODIFY: Flask application now permanently stores data in database file and in json files